+++ NEWS.md
@@ -9,6 +9,8 @@
 * Update `mlaps()` hospital query
 * New option to derive ICU entries as a clinical outcome with customizable time windows in `icu_entry()`
 * Small bug fix in `disability()`
+* New function `convert_dt()` to handle date-time variables and show missing/invalid entries.
+* Enhancement of `Rgemini` functions to allow for different date-time inputs.
 * Updated `Rgemini:::check_input()` to return error for `odbc` connections 
 * Updated unit test and small bug fix in `frailty_score()`
 * Adds `hospitalizations_last_n_days()` to calculate the number of previous hospitalizations in a given window.